Add comparisons function and sanitize data after scrap

+++ src/utils/files.ts
@@ -1,14 +1,14 @@
-import { existsSync, mkdirSync, readFileSync, rmSync, writeFileSync } from "fs";
+import { existsSync, mkdirSync, readdirSync, readFileSync, rmSync, writeFileSync } from "fs";
 import { join, dirname } from "path";
 
 export const ensureDirectory = (path: string) => existsSync(path) || mkdirSync(path, { recursive: true });
 
-export const toJsonFile = async (data: string, fileName: string): Promise<void> => {
+export const toJsonFile = async (data: string, fileName: string, overwrite = false): Promise<void> => {
     const dir = join(process.cwd(), 'src', 'data', 'autoscout');
     ensureDirectory(dir);
     const filePath = join(dir, `${fileName}.json`);
     dirname(filePath) && ensureDirectory(dirname(filePath));
-    if (existsSync(filePath)) {
+    if (!overwrite && existsSync(filePath)) {
         const existentData = JSON.parse(readFileSync(filePath, 'utf8'));
         return writeFileSync(filePath, JSON.stringify([...existentData, ...JSON.parse(data)]));
     }
@@ -53,4 +53,20 @@ export const readJsonFile = (fileName: string): any => {
     const filePath = join(dir, `${fileName}.json`);
     const json = readFileSync(filePath, 'utf8');
     return JSON.parse(json);
+}
+
+export const readAllDirFiles = (dir: string): any[] => {
+    const dirPath = join(process.cwd(), 'src', 'data', 'autoscout', dir);
+    const files = readdirSync(dirPath).map((file) => readJsonFile(`${dir}/${file.split('.json')[0]}`));
+    return files;
+}
+
+export const writeAllDirFiles = (dir: string, newData: any[][]): any[] => {
+    const dirPath = join(process.cwd(), 'src', 'data', 'autoscout', dir);
+    const files = readdirSync(dirPath).map((file, index) => toJsonFile(
+        JSON.stringify(newData[index]),
+        `${dir}${file.split('.json')[0]}`,
+        true
+    ));
+    return files;
 }
